Add Dashboard component, routes, and dependencies
Introduced a new `Dashboard` component with various charts to display evaluation data. Updated the application routes in `App.js` to include the `Dashboard` and a new `Login` page. Several new dependencies, such as `recharts` and `jwt-decode`, were added to `package.json` and `package-lock.json`. Also replaced the `NavigationHeader` component with a more comprehensive `Navbar`. Added a `.env` to the `.gitignore` file for environment variable management.